move 3.11+ locations table parsing to opcodes
