+++ xdis/opcodes/opcode_311.py
@@ -259,3 +259,113 @@ opcode_extended_fmt = opcode_extended_fmt311
 
 update_pj3(globals(), loc)
 finalize_opcodes(loc)
+
+def parse_location_entries(location_bytes, first_line):
+    """
+    Parses the locations table described in: https://github.com/python/cpython/blob/3.11/Objects/locations.md
+    The locations table replaced the line number table starting in 3.11
+    """
+    def starts_new_entry(b): return bool(b & 0b10000000) # bit 7 is set
+    def extract_code(b): return (b & 0b01111000) >> 3 # extracts bits 3-6
+    def extract_length(b): return (b & 0b00000111) + 1 # extracts bit 0-2
+
+    def iter_location_codes(loc_bytes):
+        if len(loc_bytes) == 0:
+            return []
+        
+        iter_locs = iter(loc_bytes)
+        entry_codes = [next(iter_locs)]
+
+        for b in iter_locs:
+            if starts_new_entry(b):
+                yield entry_codes
+                entry_codes = [b]
+            else:
+                entry_codes.append(b)
+
+    def iter_varints(varint_bytes):
+        if len(varint_bytes) == 0:
+            return []
+        
+        def has_next_byte(b): return bool(b & 0b0100_0000) # has bit 6 set
+        def get_value(b): return (b & 0b00111111) # extracts bits 0-5
+
+        iter_varint_bytes = iter(varint_bytes)
+
+        current_value = 0
+        shift_amt = 0
+
+        for b in iter_varint_bytes:
+            current_value += get_value(b) << shift_amt
+            if has_next_byte(b):
+                shift_amt += 6
+            else:
+                yield current_value
+                current_value = 0
+                shift_amt = 0
+
+    def decode_signed_varint(s): return -(s>>1) if s & 1 else (s>>1)
+
+    entries = [] # tuples of (code units, start line, end line, start column, end column)
+
+    last_line = first_line
+
+    for location_codes in iter_location_codes(location_bytes):
+        first_byte = location_codes[0]
+        location_length = extract_length(first_byte)
+        code = extract_code(first_byte)
+
+        if code <= 9: # short form
+            start_line = last_line
+            end_line = start_line
+            second_byte = location_codes[1]
+            start_column = (code*8) + ((second_byte>>4)&7)
+            end_column = start_column + (second_byte&15)
+        elif code <= 12: # one line form
+            start_line = last_line + code - 10
+            end_line = start_line
+            start_column = location_codes[1]
+            end_column = location_codes[2]
+        elif code == 13: # no column info
+            (start_line_delta,) = iter_varints(location_codes[1:])
+            start_line = last_line + decode_signed_varint(start_line_delta)
+            end_line = start_line
+            start_column = None
+            end_column = None
+        elif code == 14: # long form
+            (start_line_delta, end_line_delta, start_column, end_column) = iter_varints(location_codes[1:])
+            start_line = last_line + decode_signed_varint(start_line_delta)
+            end_line = start_line + end_line_delta
+        else: # code == 15, no location
+            start_line = None
+            end_line = None
+            start_column = None
+            end_column = None
+
+        entries.append((location_length, start_line, end_line, start_column, end_column))
+
+        last_line = start_line if start_line is not None else last_line
+
+    return entries
+
+def findlinestarts(code, dup_lines=False):
+    """Find the offsets in a byte code which are start of lines in the source.
+
+    Generate pairs (offset, lineno) as described in Python/compile.c.
+    """
+    location_table_entries = parse_location_entries(code.co_lnotab, code.co_firstlineno)
+
+    current_offset = 0
+    previous_line = None
+
+    for offset_delta, start_line, _, _, _ in location_table_entries:
+
+        # emit an entry when the line changes; start_line==None means there is no line information
+        if start_line is not None and (start_line != previous_line or dup_lines):
+            yield current_offset, start_line
+            previous_line = start_line
+    
+        current_offset += 2*offset_delta
+
+update_pj3(globals(), loc)
+finalize_opcodes(loc)
